requirements.txt
flask
flask-cors
numpy
pillow
keras==2.12.0
tensorflow==2.12.0